+++ src/something_projectile.cpp
@@ -88,7 +88,7 @@ void update_projectiles(float dt)
         case Projectile_State::Active: {
             update_animat(&projectiles[i].active_animat, dt);
             projectiles[i].pos += projectiles[i].vel;
-            const auto projectile_tile = projectiles[i].pos / (float) TILE_SIZE;
+            const auto projectile_tile = projectiles[i].pos / TILE_SIZE;
             if (!is_tile_empty(vec_cast<int>(projectile_tile))
                 || !rect_contains_vec2(LEVEL_BOUNDARY, projectiles[i].pos)) {
                 projectiles[i].state = Projectile_State::Poof;
